+++ pkg/db/ent/application.go
@@ -29,11 +29,11 @@ type Application struct {
 	// ApplicationLogo holds the value of the "application_logo" field.
 	ApplicationLogo string `json:"application_logo,omitempty"`
 	// CreateAt holds the value of the "create_at" field.
-	CreateAt int64 `json:"create_at,omitempty"`
+	CreateAt uint32 `json:"create_at,omitempty"`
 	// UpdateAt holds the value of the "update_at" field.
-	UpdateAt int64 `json:"update_at,omitempty"`
+	UpdateAt uint32 `json:"update_at,omitempty"`
 	// DeleteAt holds the value of the "delete_at" field.
-	DeleteAt int64 `json:"delete_at,omitempty"`
+	DeleteAt uint32 `json:"delete_at,omitempty"`
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
@@ -108,19 +108,19 @@ func (a *Application) assignValues(columns []string, values []interface{}) error
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field create_at", values[i])
 			} else if value.Valid {
-				a.CreateAt = value.Int64
+				a.CreateAt = uint32(value.Int64)
 			}
 		case application.FieldUpdateAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field update_at", values[i])
 			} else if value.Valid {
-				a.UpdateAt = value.Int64
+				a.UpdateAt = uint32(value.Int64)
 			}
 		case application.FieldDeleteAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field delete_at", values[i])
 			} else if value.Valid {
-				a.DeleteAt = value.Int64
+				a.DeleteAt = uint32(value.Int64)
 			}
 		}
 	}
