+++ pkg/db/ent/applicationgroup/where.go
@@ -127,21 +127,21 @@ func Annotation(v string) predicate.ApplicationGroup {
 }
 
 // CreateAt applies equality check predicate on the "create_at" field. It's identical to CreateAtEQ.
-func CreateAt(v int64) predicate.ApplicationGroup {
+func CreateAt(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // UpdateAt applies equality check predicate on the "update_at" field. It's identical to UpdateAtEQ.
-func UpdateAt(v int64) predicate.ApplicationGroup {
+func UpdateAt(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldUpdateAt), v))
 	})
 }
 
 // DeleteAt applies equality check predicate on the "delete_at" field. It's identical to DeleteAtEQ.
-func DeleteAt(v int64) predicate.ApplicationGroup {
+func DeleteAt(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldDeleteAt), v))
 	})
@@ -696,21 +696,21 @@ func AnnotationContainsFold(v string) predicate.ApplicationGroup {
 }
 
 // CreateAtEQ applies the EQ predicate on the "create_at" field.
-func CreateAtEQ(v int64) predicate.ApplicationGroup {
+func CreateAtEQ(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtNEQ applies the NEQ predicate on the "create_at" field.
-func CreateAtNEQ(v int64) predicate.ApplicationGroup {
+func CreateAtNEQ(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtIn applies the In predicate on the "create_at" field.
-func CreateAtIn(vs ...int64) predicate.ApplicationGroup {
+func CreateAtIn(vs ...uint32) predicate.ApplicationGroup {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -727,7 +727,7 @@ func CreateAtIn(vs ...int64) predicate.ApplicationGroup {
 }
 
 // CreateAtNotIn applies the NotIn predicate on the "create_at" field.
-func CreateAtNotIn(vs ...int64) predicate.ApplicationGroup {
+func CreateAtNotIn(vs ...uint32) predicate.ApplicationGroup {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -744,49 +744,49 @@ func CreateAtNotIn(vs ...int64) predicate.ApplicationGroup {
 }
 
 // CreateAtGT applies the GT predicate on the "create_at" field.
-func CreateAtGT(v int64) predicate.ApplicationGroup {
+func CreateAtGT(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtGTE applies the GTE predicate on the "create_at" field.
-func CreateAtGTE(v int64) predicate.ApplicationGroup {
+func CreateAtGTE(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtLT applies the LT predicate on the "create_at" field.
-func CreateAtLT(v int64) predicate.ApplicationGroup {
+func CreateAtLT(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldCreateAt), v))
 	})
 }
 
 // CreateAtLTE applies the LTE predicate on the "create_at" field.
-func CreateAtLTE(v int64) predicate.ApplicationGroup {
+func CreateAtLTE(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldCreateAt), v))
 	})
 }
 
 // UpdateAtEQ applies the EQ predicate on the "update_at" field.
-func UpdateAtEQ(v int64) predicate.ApplicationGroup {
+func UpdateAtEQ(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtNEQ applies the NEQ predicate on the "update_at" field.
-func UpdateAtNEQ(v int64) predicate.ApplicationGroup {
+func UpdateAtNEQ(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtIn applies the In predicate on the "update_at" field.
-func UpdateAtIn(vs ...int64) predicate.ApplicationGroup {
+func UpdateAtIn(vs ...uint32) predicate.ApplicationGroup {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -803,7 +803,7 @@ func UpdateAtIn(vs ...int64) predicate.ApplicationGroup {
 }
 
 // UpdateAtNotIn applies the NotIn predicate on the "update_at" field.
-func UpdateAtNotIn(vs ...int64) predicate.ApplicationGroup {
+func UpdateAtNotIn(vs ...uint32) predicate.ApplicationGroup {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -820,49 +820,49 @@ func UpdateAtNotIn(vs ...int64) predicate.ApplicationGroup {
 }
 
 // UpdateAtGT applies the GT predicate on the "update_at" field.
-func UpdateAtGT(v int64) predicate.ApplicationGroup {
+func UpdateAtGT(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtGTE applies the GTE predicate on the "update_at" field.
-func UpdateAtGTE(v int64) predicate.ApplicationGroup {
+func UpdateAtGTE(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtLT applies the LT predicate on the "update_at" field.
-func UpdateAtLT(v int64) predicate.ApplicationGroup {
+func UpdateAtLT(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldUpdateAt), v))
 	})
 }
 
 // UpdateAtLTE applies the LTE predicate on the "update_at" field.
-func UpdateAtLTE(v int64) predicate.ApplicationGroup {
+func UpdateAtLTE(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldUpdateAt), v))
 	})
 }
 
 // DeleteAtEQ applies the EQ predicate on the "delete_at" field.
-func DeleteAtEQ(v int64) predicate.ApplicationGroup {
+func DeleteAtEQ(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtNEQ applies the NEQ predicate on the "delete_at" field.
-func DeleteAtNEQ(v int64) predicate.ApplicationGroup {
+func DeleteAtNEQ(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtIn applies the In predicate on the "delete_at" field.
-func DeleteAtIn(vs ...int64) predicate.ApplicationGroup {
+func DeleteAtIn(vs ...uint32) predicate.ApplicationGroup {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -879,7 +879,7 @@ func DeleteAtIn(vs ...int64) predicate.ApplicationGroup {
 }
 
 // DeleteAtNotIn applies the NotIn predicate on the "delete_at" field.
-func DeleteAtNotIn(vs ...int64) predicate.ApplicationGroup {
+func DeleteAtNotIn(vs ...uint32) predicate.ApplicationGroup {
 	v := make([]interface{}, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -896,28 +896,28 @@ func DeleteAtNotIn(vs ...int64) predicate.ApplicationGroup {
 }
 
 // DeleteAtGT applies the GT predicate on the "delete_at" field.
-func DeleteAtGT(v int64) predicate.ApplicationGroup {
+func DeleteAtGT(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtGTE applies the GTE predicate on the "delete_at" field.
-func DeleteAtGTE(v int64) predicate.ApplicationGroup {
+func DeleteAtGTE(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtLT applies the LT predicate on the "delete_at" field.
-func DeleteAtLT(v int64) predicate.ApplicationGroup {
+func DeleteAtLT(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldDeleteAt), v))
 	})
 }
 
 // DeleteAtLTE applies the LTE predicate on the "delete_at" field.
-func DeleteAtLTE(v int64) predicate.ApplicationGroup {
+func DeleteAtLTE(v uint32) predicate.ApplicationGroup {
 	return predicate.ApplicationGroup(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldDeleteAt), v))
 	})
