+++ pkg/db/ent/applicationgroupuser.go
@@ -25,9 +25,9 @@ type ApplicationGroupUser struct {
 	// Annotation holds the value of the "annotation" field.
 	Annotation string `json:"annotation,omitempty"`
 	// CreateAt holds the value of the "create_at" field.
-	CreateAt int64 `json:"create_at,omitempty"`
+	CreateAt uint32 `json:"create_at,omitempty"`
 	// DeleteAt holds the value of the "delete_at" field.
-	DeleteAt int64 `json:"delete_at,omitempty"`
+	DeleteAt uint32 `json:"delete_at,omitempty"`
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
@@ -90,13 +90,13 @@ func (agu *ApplicationGroupUser) assignValues(columns []string, values []interfa
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field create_at", values[i])
 			} else if value.Valid {
-				agu.CreateAt = value.Int64
+				agu.CreateAt = uint32(value.Int64)
 			}
 		case applicationgroupuser.FieldDeleteAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field delete_at", values[i])
 			} else if value.Valid {
-				agu.DeleteAt = value.Int64
+				agu.DeleteAt = uint32(value.Int64)
 			}
 		}
 	}
